Update lab1.rst
Reversed order of two examples (running a command without accessing the shell, and running a command when the pod includes more than one container). I believe this is clearer to the student. I also added language to let the student know that explicitly specifying the container can also be done in single-container pods, but is optional.