Update lab1.rst
In the actual lab (from the "AppWorld 2025 - NGINX Plus 301: CI/CD with NGINX+ (and Gitlab) (version 1)" blueprint), the www.appster.com.conf file contains the line,
# limit_req $binary_remote_addr zone=limit_me:1m rate=10r/s;

In this version of the blueprint, It is initially set to 10, not 10000. This PR fixes the text that suggests otherwise (and would be a potential source of confusion for the students).